Renamed copy dir and their classes into worker
