--- app/src/main/java/org/qp/android/model/copy/CopyException.java
@@ -1,8 +1,0 @@
-package org.qp.android.model.copy;
-
-public class CopyException extends RuntimeException {
-
-    public CopyException(String message) {
-        super(message);
-    }
-}
