+++ app/src/main/java/org/qp/android/model/workers/WorkerException.java
@@ -1,0 +1,8 @@
+package org.qp.android.model.workers;
+
+public class WorkerException extends RuntimeException {
+
+    public WorkerException(String message) {
+        super(message);
+    }
+}
